refactor(protocol): remove server contract shims
