+++ packages/schema/src/project-copy.ts
@@ -1,14 +1,14 @@
 export * as ProjectCopy from "./project-copy"
 
 import { Schema } from "effect"
-import { Project } from "./project"
+import { ProjectID } from "./project-id"
 import { AbsolutePath } from "./schema"
 
 export const StrategyID = Schema.Trim.pipe(Schema.check(Schema.isNonEmpty()), Schema.brand("ProjectCopy.StrategyID"))
 export type StrategyID = typeof StrategyID.Type
 
 export const CreateInput = Schema.Struct({
-  projectID: Project.ID,
+  projectID: ProjectID,
   strategy: StrategyID,
   sourceDirectory: AbsolutePath,
   directory: AbsolutePath,
@@ -17,7 +17,7 @@ export const CreateInput = Schema.Struct({
 export type CreateInput = typeof CreateInput.Type
 
 export const RemoveInput = Schema.Struct({
-  projectID: Project.ID,
+  projectID: ProjectID,
   directory: AbsolutePath,
   force: Schema.Boolean,
 }).annotate({ identifier: "ProjectCopy.RemoveInput" })
