--- packages/server/src/api.ts
@@ -1,1 +1,0 @@
-export { Api, makeApi } from "@opencode-ai/protocol/api"
