--- packages/server/src/errors.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/errors"
