--- packages/server/src/groups/agent.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/groups/agent"
