--- packages/server/src/groups/command.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/groups/command"
