--- packages/server/src/groups/credential.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/groups/credential"
