--- packages/server/src/groups/event.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/groups/event"
