--- packages/server/src/groups/fs.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/groups/fs"
