--- packages/server/src/groups/health.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/groups/health"
