--- packages/server/src/groups/integration.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/groups/integration"
