--- packages/server/src/groups/message.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/groups/message"
