--- packages/server/src/groups/model.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/groups/model"
