--- packages/server/src/groups/permission.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/groups/permission"
