--- packages/server/src/groups/project-copy.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/groups/project-copy"
