--- packages/server/src/groups/provider.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/groups/provider"
