--- packages/server/src/groups/question.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/groups/question"
