--- packages/server/src/groups/reference.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/groups/reference"
