--- packages/server/src/groups/session.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/groups/session"
