--- packages/server/src/groups/skill.ts
@@ -1,1 +1,0 @@
-export * from "@opencode-ai/protocol/groups/skill"
