+++ packages/server/src/handlers/agent.ts
@@ -2,7 +2,7 @@ import { AgentV2 } from "@opencode-ai/core/agent"
 import { Effect } from "effect"
 import { HttpApiBuilder } from "effect/unstable/httpapi"
 import { Api } from "@opencode-ai/protocol/api"
-import { response } from "../groups/location"
+import { response } from "../location"
 
 export const AgentHandler = HttpApiBuilder.group(Api, "server.agent", (handlers) =>
   handlers.handle("agent.list", () =>
