+++ packages/server/src/handlers/command.ts
@@ -1,7 +1,7 @@
 import { CommandV2 } from "@opencode-ai/core/command"
 import { HttpApiBuilder } from "effect/unstable/httpapi"
 import { Api } from "@opencode-ai/protocol/api"
-import { response } from "../groups/location"
+import { response } from "../location"
 
 export const CommandHandler = HttpApiBuilder.group(Api, "server.command", (handlers) =>
   handlers.handle("command.list", () => response(CommandV2.Service.use((command) => command.list()))),
