+++ packages/server/src/handlers/model.ts
@@ -2,7 +2,7 @@ import { Catalog } from "@opencode-ai/core/catalog"
 import { Effect } from "effect"
 import { HttpApiBuilder } from "effect/unstable/httpapi"
 import { Api } from "@opencode-ai/protocol/api"
-import { response } from "../groups/location"
+import { response } from "../location"
 
 export const ModelHandler = HttpApiBuilder.group(Api, "server.model", (handlers) =>
   Effect.gen(function* () {
