+++ packages/server/src/handlers/reference.ts
@@ -1,7 +1,7 @@
 import { Reference } from "@opencode-ai/core/reference"
 import { HttpApiBuilder } from "effect/unstable/httpapi"
 import { Api } from "@opencode-ai/protocol/api"
-import { response } from "../groups/location"
+import { response } from "../location"
 
 export const ReferenceHandler = HttpApiBuilder.group(Api, "server.reference", (handlers) =>
   handlers.handle("reference.list", () => response(Reference.Service.use((reference) => reference.list()))),
