+++ packages/server/src/handlers/skill.ts
@@ -1,7 +1,7 @@
 import { SkillV2 } from "@opencode-ai/core/skill"
 import { HttpApiBuilder } from "effect/unstable/httpapi"
 import { Api } from "@opencode-ai/protocol/api"
-import { response } from "../groups/location"
+import { response } from "../location"
 
 export const SkillHandler = HttpApiBuilder.group(Api, "server.skill", (handlers) =>
   handlers.handle("skill.list", () => response(SkillV2.Service.use((skill) => skill.list()))),
